+++ src/pages/Detail/History.jsx
@@ -5,7 +5,7 @@ import styles from './History.module.css';
 import useModal from '../../hooks/useModal';
 import OverlayPortal from '../../components/common/Overlay/OverlayPortal';
 import Overlay from '../../components/common/Overlay/Overlay';
-import ModalPortal from '../../components/common/Modal/ModalProtal';
+import ModalPortal from '../../components/common/Modal/ModalPortal';
 import ConfirmModal from '../../components/common/Modal/ConfirmModal';
 import SetHistoryIcon from '../../components/common/icons/SetHistoryIcon';
 import useClassDetail from '../../hooks/useClassDetail';
